Ensuring proper error handling on backend
Made sure all database calls (where it makes sense) or other api calls that are awaited upon are contained within try/catch blocks. This is to ensure proper error handling to prevent the entire application from crashing when the database is down. That being said, when the database is down next to nothing can be done on the site so it will just display errors.