Cart loading update and db error handling
Updated the UI for when a cart is being loaded to only show a message that the cart is empty after attempting to load the cart (Rather than while loading the cart). Also updated server side DB query methods to return errors to client if the database cannot be reached, as it doesnt throw an error but instead returns empty query results.