Fixed home page object intake
Had dont a really poor system that added a bunch of extra steps and turned objects from database into a significantly less readable array system. Realized this is very backwards and fixed the homeBody and index.js files today, but navBar isnt so the cart doesn't work alongside it. Going to fix the navBar's cart UI as well as the other relevant pages to fix this problem (tomorrow).