Entirely Removed weird array structuring of listings
Listings objects from the database were being turned into arrays to be populated throughout the page which added a lot of complexity (that did next to nothing) and also made the code unreadable. used this system when i started this project back in July 2021 and it was very  bad and silly as I can see now. It has been completely overhauled to utilize the listings object as defined within the database for a significant reduction in complexity and increase in readability. Still unsure if I want to overhaul how the cart is stored as well as i currently only store the object ID of the listing as well as the quantity and then populate on the frontend. I could store the entire object in the cart, but that means I have to validate post requests or else someone could add random items to it that aren't a part of the application.