Reformatted Orders to contain full objects
Reformatted the Orders as they are saved in the database to be populated with full objects rather than the miniaturized arrays I save carts as. This saves a lot of work on the frontend.  Added checkout functionality on backend so checking out now wipes carts and assigns it as an "Order" with a timestamp, linked account, and cart if the user is logged into an account. If it is a guest, the order is currently thrown out as there isn't a real checkout since the site is a mockup, not a functional store. No UI for seeing your orders even though they are saved, but see line 657 for an explanation on how you can see these orders before that is integrated.