Added desktop account page and removed excess logs
Added an account page that is somewhat done (decent amount of UI work to be done, but a rough model is finished and data is successfully retrieved.) Webpack config done for this, and partitioning of CSS classes has been done so i know what pieces of the model I'll be working with. Removed some excess logging on backend and frontend. Orders for guests (users who are not logged into accounts) are currently deleted, but I may add a guest checkout where a user can input an email for their order to be retrieved through a guest orders interface.